+++ src/main/scala/dev/template/spark/CovidDataHivePartitioner.scala
@@ -7,19 +7,24 @@ import org.apache.spark.sql.Row
 import org.apache.spark.sql.SaveMode
 import org.apache.spark.sql.SparkSession
 
-object CovidDataPartitioner
+object CovidDataHivePartitioner
     extends App
     with SparkSessionWrapper
     with Reader
     with Writer
     with Logger {
 
-  def writeParquet(spark: SparkSession, file: String, outputPath: String): Unit = {
+  def writeParquet(
+    spark: SparkSession,
+    file: String,
+    outputPath: String,
+    schema_location: String
+  ): Unit = {
     val covidData: Dataset[Row] = readCsv(spark).csv(file).repartition(1)
     log.info(covidData.printSchema())
     covidData.createOrReplaceTempView("covid")
 
-    val groupedView = sqlContext
+    val groupedView = spark.sqlContext
       .sql("""
              | select
              |  cast(to_date(date, "yyyy-MM-dd") as String) as reported_date,
@@ -32,10 +37,36 @@ object CovidDataPartitioner
              |""".stripMargin)
       .cache()
 
-    log.info(covidData.printSchema())
+    val createSchema = spark.sql(s""" create schema if not exists public_data
+                                    |LOCATION "$schema_location"
+                                    |""".stripMargin)
+    createSchema.show()
 
-    writeParquet(groupedView, outputPath, SaveMode.Overwrite, Some("reported_date"))
+    log.info(s"Schema Created ${createSchema.printSchema()}")
 
+    val tableCreated = spark.sql(s""" create external table if not exists public_data.covid
+                                    | (
+                                    | reported_date date,
+                                    | county string,
+                                    | state string,
+                                    | fips int,
+                                    | cases int,
+                                    | deaths int
+                                    |)
+                                    |  partitioned by (reported_date)
+                                    |  Stored as parquet
+                                    |LOCATION "$outputPath"
+                                    |""".stripMargin)
+
+    log.info(s"Table created $tableCreated.printSchema() ")
+
+    writeParquetAsTable(
+      groupedView,
+      SaveMode.Overwrite,
+      "poc.covid",
+      path = outputPath,
+      Some("reported_date")
+    )
   }
 
   if (args.length == 0) {
@@ -57,9 +88,10 @@ object CovidDataPartitioner
 
   val inputFilePath = args(0)
   val outputPath: String = args(1)
+  val hive_schema_path: String = args(2)
   log.info("Input path " + inputFilePath)
   log.info("Output path " + outputPath)
 
-  writeParquet(spark, inputFilePath, outputPath)
+  writeParquet(spark, inputFilePath, outputPath, hive_schema_path)
 
 }