+++ src/test/scala/dev/template/spark/CovidDataPartitionerTest.scala
@@ -1,6 +1,8 @@
 package dev.template.spark
 
-import org.apache.hadoop.fs.{FileSystem, Path, PathFilter}
+import org.apache.hadoop.fs.FileSystem
+import org.apache.hadoop.fs.Path
+import org.apache.hadoop.fs.PathFilter
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.runner.RunWith
 import org.scalatest.funspec.AnyFunSpec
@@ -11,9 +13,13 @@ class CovidDataPartitionerTest extends AnyFunSpec with SparkSessionTestWrapper {
 
   describe("should write data partitioned by reported date ") {
     val file = ClassLoader.getSystemResource("us-counties-recent.csv").getFile
-    val outputPath = "/tmp/covid-data/us_covid_data/"
+    val outputPath = ClassLoader.getSystemResource("").getFile + "/us-counties-output/"
     CovidDataPartitioner.writeParquet(spark, file, outputPath)
     it("returns 3257 rows for reported_date=2023-03-23") {
+      val outputPath = ClassLoader.getSystemResource("").getFile + "/us-counties-output/"
+
+      println(outputPath)
+
       def read = spark.read.parquet(outputPath + "reported_date=2023-03-23")
 
       assertEquals(3257, read.count())
@@ -23,21 +29,22 @@ class CovidDataPartitionerTest extends AnyFunSpec with SparkSessionTestWrapper {
 
       val fs = FileSystem.get(spark.sparkContext.hadoopConfiguration)
       val reportedDateFolderCount = fs
-        .listStatus(new Path(outputPath),
-                    new PathFilter {
-                      override def accept(path: Path): Boolean =
-                        path.getName.contains("reported_date")
-                    })
+        .listStatus(
+          new Path(outputPath),
+          new PathFilter {
+            override def accept(path: Path): Boolean =
+              path.getName.contains("reported_date")
+          }
+        )
         .length
 
       assertEquals(30, reportedDateFolderCount)
 
-      /**
-       * the below CTAS fails when running in java 8 with the error message. Leaving it here for
-       * reference and replacing the test to get count with FileSystem API
+      /** the below CTAS fails when running in java 8 with the error message. Leaving it here for reference and
+       * replacing the test to get count with FileSystem API
        *
-       * A fatal error has been detected by the Java Runtime Environment: # # SIGSEGV (0xb) at
-       * pc=0x000000010c0d604e, pid=65727, tid=0x0000000000002703
+       * A fatal error has been detected by the Java Runtime Environment: # # SIGSEGV (0xb) at pc=0x000000010c0d604e,
+       * pid=65727, tid=0x0000000000002703
        */
 
       //      def read =
@@ -51,4 +58,5 @@ class CovidDataPartitionerTest extends AnyFunSpec with SparkSessionTestWrapper {
       //      assertEquals(30, part.count())
     }
   }
+
 }