+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-
 # Change Log
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 First release version of the project
 
 ### Added
+
 - support for gradle 7.6, java 11, spark 3.4.1, scala 2.12 and delta 2.4.0
 - support to read file:// and s3://
 - Added CHANGELOG.md
@@ -18,3 +19,14 @@ First release version of the project
 - Added tests for the above classes
 - Added scalaCodeCoverage
 - Added spotless, checkstyle, and spotbugs
+
+## [1.0.0] - 2023-02-03
+
+Run this application on Google Cloud Dataproc
+
+### Added
+
+- Added word count examples and test
+- Added GCS libraries and tests
+- Added abris
+- Added Hive sample app with public covid data